+++ src/com/hueemulator/gui/About.java
@@ -7,15 +7,17 @@ import javax.swing.JFrame;
 public class About extends JFrame{
     private JEditorPane aboutPane;
 
+    private static String VERSION = "v0.4";
+    
     public About() {
 
-        String text = "<h2>Hue Emulator v0.2</h2>Programmed By:<br> Stevey O";
+        String text = "<h2>Hue Emulator " + VERSION + "</h2><p>Programmed By:&nbsp;&nbsp;<b>SteveyO</b>. <p>See <a href=\"http://steveyo.github.io/Hue-Emulator/\">http://steveyo.github.io/Hue-Emulator/</a> for more details.</p>";
 
         aboutPane = new JEditorPane();
         aboutPane.setContentType("text/html");
         aboutPane.setText(text);
 
-        setSize(300, 200);
+        setSize(400, 210);
         aboutPane.setVisible(true);
         add(aboutPane);
     }